// src/app/models/detalhes-fato.ts
export class DetalhesFato {
    codigo: any;
    codigoTipo:any;
    codigoLocal:any;
    data: any;
    hora: any;
    descricao: any;
    uf: any;
    tipo_guia: string;

    municipio: any;
    nome_municipio: any;

    bairro: any;
    nome_bairro: any;

    tipo_local: any;
    nome_tipo_local: any;

    tipo_logradouro: any;
    nome_logradouro: any;
    complemento: any;
    cep: any;
    natureza_do_fato: any;

    tipificacao: any;
    nome_tipificacao: any;

    descricao_local:any;

    descricao_da_suspeita: any;
    numguia: any;
    numocur: any;
    rascunho: any;
    autoridade: any;
    usuario: any;

    constructor() {

    }
}